refactor(subagent,feedback,sanitizer): deduplicate hook env, fix visibility, add non_exhaustive
Extract make_base_hook_env + TOOL_ARGS_JSON_LIMIT to zeph-subagent/hooks.rs,
removing byte-for-byte duplication between agent_loop.rs and tier_loop.rs (#4015).

Make JudgeDetector.call_times private; rewrite rate-limiter test to use
tokio::time::pause()/advance() instead of direct field manipulation (#3988).

Add #[non_exhaustive] to ContentSourceKind and ContentTrustLevel in
zeph-sanitizer to prevent downstream match exhaustion on new variants (#3932).

Closes #4015
Closes #3988
Closes #3932
