+++ CHANGELOG.md
@@ -16,6 +16,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
   `context.trajectory_hints`, `context.tree_memory`) to the three async context fetchers in the
   assembler (closes #3984).
 
+### Refactored
+
+- `zeph-subagent`: extracted `make_base_hook_env` and `TOOL_ARGS_JSON_LIMIT` into
+  `zeph-subagent::hooks`, eliminating duplicate env-building logic between subagent and
+  core tier-loop hook dispatch (closes #4015).
+- `zeph-agent-feedback`: `JudgeDetector::call_times` is now a private field; switched
+  `VecDeque<Instant>` from `std::time::Instant` to `tokio::time::Instant` to enable
+  deterministic time control in tests (closes #3988).
+- `zeph-sanitizer`: `ContentTrustLevel` and `ContentSourceKind` are now `#[non_exhaustive]`
+  to allow adding variants without breaking external exhaustive matches (closes #3932).
+
 ### Changed
 
 - `zeph-commands`: `GoalCommand::handle` now propagates errors via `?` instead of silently converting
